passes all tests and verify stage

+++ src/main/java/pl/coderstrust/database/file/InFileDatabase.java
@@ -1,41 +1,69 @@
 package pl.coderstrust.database.file;
 
 import pl.coderstrust.database.Database;
+import pl.coderstrust.database.DbException;
+import pl.coderstrust.database.ExceptionMsg;
 import pl.coderstrust.database.ObjectMapperHelper;
 import pl.coderstrust.model.Invoice;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.stream.Collectors;
 
 public class InFileDatabase implements Database {
 
+  private static final int FIRST_ID = 0;
+  private static final int ID_INCREMENT = 1;
+
   private FileHelper fileHelper;
   private ObjectMapperHelper mapper;
   private HashSet<Long> savedIds;
+  long lastId;
 
   public InFileDatabase() {
     mapper = new ObjectMapperHelper();
     fileHelper = new FileHelper();
-    savedIds = getSavedIds();
+    savedIds = getIdsFromDbFile();
   }
 
   @Override
-  public void addInvoice(Invoice invoice) {
+  public long addInvoice(Invoice invoice) {
+    invoice.setId(getNextId());
     fileHelper.addLine(mapper.toJson(invoice));
     savedIds.add(invoice.getId());
+    return invoice.getId();
+  }
+
+  private long getNextId() {
+    if (savedIds.isEmpty()) {
+      return FIRST_ID;
+    } else {
+      return Collections.max(savedIds) + ID_INCREMENT;
+    }
   }
 
   @Override
   public void deleteInvoice(long systemId) {
-    fileHelper.deleteLine(idToLineKey(systemId));
-    savedIds.remove(systemId);
+    if (!idExist(systemId)) {
+      throw new DbException(ExceptionMsg.INVOICE_NOT_EXIST);
+      //TODO change to logging;
+    } else {
+      fileHelper.deleteLine(idToLineKey(systemId));
+      savedIds.remove(systemId);
+    }
   }
 
   @Override
   public Invoice getInvoiceById(long systemId) {
-    String jsonInvoice = fileHelper.getLine(idToLineKey(systemId));
-    return mapper.toInvoice(jsonInvoice);
+    if (!idExist(systemId)) {
+      throw new DbException(ExceptionMsg.INVOICE_NOT_EXIST);
+      //TODO change to logging;
+    } else {
+
+      String jsonInvoice = fileHelper.getLine(idToLineKey(systemId));
+      return mapper.toInvoice(jsonInvoice);
+    }
   }
 
   String idToLineKey(long systemId) {
@@ -45,7 +73,8 @@ public class InFileDatabase implements Database {
   @Override
   public void updateInvoice(Invoice invoice) {
     deleteInvoice(invoice.getId());
-    addInvoice(invoice);
+    fileHelper.addLine(mapper.toJson(invoice));
+    savedIds.add(invoice.getId());
   }
 
   @Override
@@ -60,7 +89,7 @@ public class InFileDatabase implements Database {
     return savedIds.contains(id);
   }
 
-  private HashSet<Long> getSavedIds() {
+  private HashSet<Long> getIdsFromDbFile() {
     return getInvoices().stream()
         .map(invoice -> invoice.getId())
         .collect(Collectors.toCollection(HashSet::new));
